requirements.txt
python==3.11.8
pygame==2.5.2 
numpy==1.26.4
pyyaml==6.0.1 